+++ netsharp-basebiz-web/src/main/java/org/netsharp/organization/controller/authorization/operation/OperationAuthorizationPart.java
@@ -244,7 +244,7 @@ public class OperationAuthorizationPart extends TreegridPart {
 			ap = new AuthorizationPrincipal();
 			{
 				ap.setEntityState(EntityState.New);
-				ap.setPrincipalId(Integer.valueOf(principalId.toString()));
+				ap.setPrincipalId(Long.valueOf(principalId.toString()));
 				ap.setPrincipalType(PrincipalType.POST);
 			}
 		}
